requirements.txt
discord-py-slash-command